Remove delegate::bind with pointer argument
Delegate supports 2 overloads for binding member functions;
one that supports a raw pointer to the entity being bound,
and the other that accepts a reference. There is no distinction
between the two, but there is an implicit requirement that the
pointer always not be null, since otherwise invoking the delegate
would lead to undefined behaviour (invoking null pointer).

This is considered an API bug. To fix this, the pointer overloads
have been removed in favor of just the reference ones, which
guarantee non-nullability.

Signed-off-by: Matthew Rodusek <matthew.rodusek@gmail.com>