Rename target 'bit_stl' to 'stl' in CMakeLists (#18)
The target name 'bit_stl', though more unique and preferable,
seems to conflict with CMake's built-in target exporting feature.
When exporting a target (so that it can be used as an `IMPORTED`
target in a consuming library), a namespace *prefix* can be
specified that gets prepended to the target name with a scope-
resolution (e.g. a target with namespace of <prefix> becomes
<prefix>::target). This lead bit_stl to be exported with
bit::bit_stl -- which conflicts directly with the alias that is
used for simple imports (bit::stl vs bit::bit_stl)

This may just be a temporary solution until a better one is found, but
so far it appears that CMake is unable to rename a target while
exporting it, and exporting only an aggregate target requires each
dependent target to also be exported (resulting in multiple targets).

This is fixes what is otherwise considered a bug, since the expected
dependency identifier should be bit::stl whether the library is
recursively picked up, or just found with the FindBit module.