Cleanup miscellaneous implementation details (#22)
* Cleanup 'delegate' implementation

The 'delegate' class now makes use of the 'function_pointer', and
'member_function_pointer' utilities.

* Fix 'exclusive_ptr' constructors

The constructors accepting 'nullptr_t' and a deleter or allocator
should create an empty control block which is null, but not valueless
in order to be consistent with std::shared_ptr.

* SFINAE-disable nth_type when n > sizeof...(Types)

This now provides a template specialization where n is larger than
sizeof...(Types), which results in ::type not being defined. This
allows SFINAE-disabling functionality when out-of-range, leading to
easier detection

* Use inheritance for [bool_/size_]constant

Both bool_constant and size_constant have now been updated to use
inheritance, rather than template aliasing to define the respective
types. Behaviourally this will be _mostly_ identical to aliases, with
the exception that `std::is_same<T,U>` will be different for queries
of `std::integral_constant` and `bool_constant`.

* Fix documentation typo in array.hpp

* Fix 'allocator_deleter' not destructing object

allocator_deleter's `operator()` function was missing a call to
'destroy' before deleting the associated object, which could lead to
memory leakage