Fix MSVC errors
The new version of MSVC decided to break expression-sfinae syntax
when used in out-of-line definitions. To avoid further problems,
the trailing decltype syntax was dropped for anything where _MSC_VER.

It is simply too volatile to trust MSVC's compilation accuracy
between versions, and so there is no sense in properly detecting the
return type (which would otherwise was added to help IDE's deduce
the return type better).