Add adl::get for tuple utilities (#35)
* Add adl namespace documentation to types

* Use adl::get for get overloads

The various tuple utility functions now use bit::stl::adl::get in order
to adl-dispatch to the appropriate get functions for tuple-like
classes

* Fix pedantic error

* Fix msvc bug make_* deduction functions

* Fix MSVC errors

The new version of MSVC decided to break expression-sfinae syntax
when used in out-of-line definitions. To avoid further problems,
the trailing decltype syntax was dropped for anything where _MSC_VER.

It is simply too volatile to trust MSVC's compilation accuracy
between versions, and so there is no sense in properly detecting the
return type (which would otherwise was added to help IDE's deduce
the return type better).