Remove 'MakeVersionHeader' cmake module
This module was not particularly useful, since the header being
generated would be better as a template file stored in the
CMake directory. As a result, this is the approach being used moving
forward since it allows for better flexibility in what is generated
into a library's individual version header.